Update Capability and yml file
