+++ tests/data_manual/demo/my_capability.yml
@@ -34,6 +34,7 @@ node_types:
       capabilities:
          some_capability:
             type: mytypes.mycapabilities.MyFeature
+            occurrences: [0, 1]
             properties:
                limit:
                   type: integer
